make Serializer/Deserializer pub
This is required when using serde_json_core with code that is
generic over the serde implementation, i.e. code that wants to
instantiate the Serializer/Deserializer, in the way done in
`de::from_slice` and `ser::to_slice`.

This also adds public getters for the Serializer.current_length
and Deserializer.index members.

Two obsolete/unused impls on the `Unreachable` have been removed.